# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel rest API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Add a new category to the store.",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/category": {
            "get": {
                "tags": [
                    "category"
                ],
                "summary": "get all category",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "category"
                ],
                "summary": "Add a new category to the store.",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "alias",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ordering",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/category/{id}": {
            "put": {
                "tags": [
                    "category"
                ],
                "summary": "Update an existing category.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "alias",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ordering",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:category",
                            "read:category"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "category"
                ],
                "summary": "Deletes a category",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:category",
                            "read:category"
                        ]
                    }
                ]
            }
        },
        "/api/subject": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "summary": "Get list of Subjects",
                "description": "Returns list of Subjects",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subject"
                ],
                "summary": "Add a new category to the store.",
                "operationId": "storeSubject",
                "requestBody": {
                    "description": "Bulk products Body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subject": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/subject/{id}": {
            "put": {
                "tags": [
                    "Subject"
                ],
                "summary": "Update an existing category.",
                "operationId": "updateSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Bulk products Body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subject": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:category",
                            "read:category"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subject"
                ],
                "summary": "Deletes a subject",
                "operationId": "destroySubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "subject id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:category",
                            "read:category"
                        ]
                    }
                ]
            }
        },
        "/api/topic": {
            "get": {
                "tags": [
                    "Topic"
                ],
                "summary": "get all topic",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "indexTopic",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Topic"
                ],
                "summary": "Add a new category to the store.",
                "operationId": "storeTopic",
                "requestBody": {
                    "description": "Bulk products Body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "topic": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/topic/{id}": {
            "put": {
                "tags": [
                    "Topic"
                ],
                "summary": "Update an existing category.",
                "operationId": "updateTopic",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "topic id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Bulk products Body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "topic": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:category",
                            "read:category"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Topic"
                ],
                "summary": "Deletes a topic",
                "operationId": "destroyTopic",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "topic id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Pet not found"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:category",
                            "read:category"
                        ]
                    }
                ]
            }
        }
    }
}